Move tests and mark client components
Reorganized `@nx/design-system` tests from `src` into a dedicated `tests/` tree and updated Vitest include paths/imports accordingly. Added `'use client'` directives to MUI-rendering brand, feedback, heading, and layout components so they run correctly in client contexts.

Also cleaned up package management metadata by deleting and ignoring the design-system `package-lock.json`, and updated `apps/www` to use `pnpm@11.13.0`.
