+++ apps/www/package.json
@@ -2,7 +2,7 @@
   "version": "3.0.6",
   "name": "www",
   "description": "NX°",
-  "packageManager": "pnpm@10.12.4",
+  "packageManager": "pnpm@11.13.0",
   "scripts": {
     "dev": "next dev -p 1999 --webpack",
     "prebuild": "pnpm --dir ../../packages/design-system build",
