+++ packages/design-system/src/components/brand/Logo.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Box, Typography } from '@mui/material';
 import type { LogoProps } from '../../types';
