+++ packages/design-system/src/components/headings/Heading.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Typography } from '@mui/material';
 import type { EyebrowProps } from '../../types';
