+++ packages/design-system/tests/components/feedback/feedback.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import { Alert, Card, Field } from './Feedback';
+import { Alert, Card, Field } from '../../../src/components/feedback/Feedback';
 
 describe('feedback components', () => {
   it('renders alert variants with optional dismiss', () => {
